+++ Changelog.md
@@ -5,10 +5,12 @@
 * Plot elevation
 * Improve `--help` text
 * Improve error message when no tracks were found
+* Add GeoJSON support
 * Allow generated website to be placed in a non-root direcotory
 * Don't re-generate photos that are already present in the output directory
-* Accept multiple input directories
+* Accept multiple input paths which can now be either directories or files
 * Show more information during the generator process (nr. photos per day, nr. skipped tiles)
+* Print total distance in the end
 
 ## v0.1.0 - 2026-05-02